+++ src/app/(sub)/token/page.tsx
@@ -1,0 +1,5 @@
+import Token from './Token';
+
+export default function page() {
+  return <Token />;
+}
